pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urbaneye"
version = "0.1.0"
description = "Autonomous Driving Perception Pipeline — CARLA Simulation, YOLOv11 Detection, ByteTrack + DeepSORT Tracking"
readme = "README.md"
license = {text = "Source Available"}
requires-python = ">=3.11"
authors = [
    {name = "Navnit Amrutharaj"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "opencv-python>=4.8",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "supervision>=0.18",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.1",
    "pytest-mock>=3.12",
    "ruff>=0.1.0",
]
training = [
    "ultralytics>=8.1",
    "albumentations>=1.3",
]
tracking = [
    "deep-sort-realtime>=1.3",
    "scipy>=1.10",
    "lap>=0.4",
    "filterpy>=1.4",
]
demo = [
    "gradio>=4.0",
    "onnxruntime>=1.16",
]
all = [
    "urbaneye[dev,training,tracking,demo]",
]

[project.urls]
Homepage = "https://github.com/ninjacode911/urbaneye"
Repository = "https://github.com/ninjacode911/urbaneye"

[tool.setuptools.packages.find]
include = ["urbaneye*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["urbaneye"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 60
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
